Cambios importantes visuales

+++ app/src/main/java/com/sabdev/froggym/MainActivity.kt
@@ -7,9 +7,7 @@ import androidx.activity.viewModels
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -118,7 +116,7 @@ fun MainScreen(authViewModel: AuthViewModel, routineViewModel: RoutineViewModel)
                 modifier = Modifier.padding(innerPadding)
             ) {
                 composable(Screen.Home.route) { HomeScreen(authViewModel = authViewModel) }
-                composable (Screen.Routines.route) {
+                composable(Screen.Routines.route) {
                     RoutinesScreen(
                         viewModel = routineViewModel,
                         onCreateRoutine = {
@@ -164,7 +162,7 @@ fun MainScreen(authViewModel: AuthViewModel, routineViewModel: RoutineViewModel)
                     val routineId = backStackEntry.arguments?.getString("routineId")
                     if (routineId != null) {
                         RoutineDetailsScreen(
-                            routineId = routineId.toInt(),
+                            routineId = routineId.toLong(),
                             viewModel = routineViewModel
                         )
                     }
