+++ app/src/main/java/com/sabdev/froggym/data/entities/RoutineExerciseCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["routineId", "exerciseId"])
 data class RoutineExerciseCrossRef(
-    val routineId: Int,
-    val exerciseId: Int
+    val routineId: Long,
+    val exerciseId: Long
 )
